requirements.txt
pytesseract==0.3.13
Pillow==11.1.0
pdf2image==1.17.0
transformers==4.39.3
openai-whisper==20231117
torch==2.2.1
